AC-921: Reconsolidated billing actions file
