--- src/actions/billing/index.js
@@ -1,2 +1,0 @@
-export * from './billing';
-export * from './account';
